requirements.txt
Flask
Flask-SQLAlchemy
FLask-Login
Flask-Admin
Flask-WTF
python-dotenv
flask-mail
pytz
weasyprint
matplotlib
pyotp
gunicorn
google-genai
pytest
locust
PyJWT
